// src/types/steam-response.ts
export interface SteamGameResponse {
	appId: number
	name: string
	playtime_forever: number
	img_icon_url: string
	has_community_visible_stats: boolean
	playtime_windows_forever: number
	playtime_mac_forever: number
	playtime_linux_forever: number
	rtime_last_played: number
	playtime_disconnected: number
}

export interface SteamOwnedGamesResponse {
	response: {
		game_count: number,
		games: SteamGameResponse[]
	}
}